seperated discretizer into own file
